+++ CHANGELOG.md
@@ -11,8 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - standalone website:
-  - classrooms page
-  - classroom create
+  - classroom (list/create/update)
 - Add new elements to the teacher VOD Dashboard :
   - Attendances from live session
 - Allow to store link between an LTI user and a "site" user